backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pandas==2.1.3
openpyxl==3.1.2
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
requests==2.31.0
httpx==0.25.2
google-generativeai==0.3.2
google-auth==2.25.2
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
email-validator==2.1.0
docling==1.7.0
docling-core==1.7.0
supabase==2.1.0
razorpay==1.3.0
numpy==1.25.2
Pillow==10.1.0
pdfplumber==0.11.0
pymupdf==1.23.8
tabulate==0.9.0
lxml==4.9.3